requirements.txt
django==5.0.4
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
phonenumbers==8.13.34
pillow==10.3.0
psycopg2==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
requests==2.32.3
django-allauth==0.63.3
django-ipware==7.0.1